Refactor poc
