+++ code/test/Templates.Test/ProjectGenerationTests.cs
@@ -66,50 +66,6 @@ namespace Microsoft.Templates.Test
         }
 
         [Theory]
-        [MemberData("GetPageAndFeatureTemplates")]
-        [Trait("Type", "OneByOneItemGeneration")]
-        public async void GenerateProjectWithIsolatedItems(string itemName, string projectType, string framework, string itemId, string language)
-        {
-            SetUpFixtureForTesting(language);
-
-            var projectTemplate = GenerationFixture.Templates.FirstOrDefault(t => t.GetTemplateType() == TemplateType.Project && t.GetProjectTypeList().Contains(projectType) && t.GetFrameworkList().Contains(framework));
-            var itemTemplate = GenerationFixture.Templates.FirstOrDefault(t => t.Identity == itemId);
-            var finalName = itemTemplate.GetDefaultName();
-            var validators = new List<Validator>
-            {
-                new ReservedNamesValidator(),
-            };
-            if (itemTemplate.GetItemNameEditable())
-            {
-                validators.Add(new DefaultNamesValidator());
-            }
-
-            finalName = Naming.Infer(finalName, validators);
-
-            var projectName = $"{projectType}{framework}{finalName}";
-
-            ProjectName = projectName;
-            ProjectPath = Path.Combine(_fixture.TestProjectsPath, projectName, projectName);
-            OutputPath = ProjectPath;
-
-            var userSelection = GenerationFixture.SetupProject(projectType, framework, language);
-
-            GenerationFixture.AddItem(userSelection, itemTemplate, GenerationFixture.GetDefaultName);
-
-            await NewProjectGenController.Instance.UnsafeGenerateProjectAsync(userSelection);
-
-            // Build solution
-            var outputPath = Path.Combine(_fixture.TestProjectsPath, projectName);
-            var result = GenerationFixture.BuildSolution(projectName, outputPath);
-
-            // Assert
-            Assert.True(result.exitCode.Equals(0), $"Solution {projectTemplate.Name} was not built successfully. {Environment.NewLine}Errors found: {GenerationFixture.GetErrorLines(result.outputFile)}.{Environment.NewLine}Please see {Path.GetFullPath(result.outputFile)} for more details.");
-
-            // Clean
-            Directory.Delete(outputPath, true);
-        }
-
-        [Theory]
         [MemberData("GetProjectTemplates")]
         [Trait("Type", "ProjectGeneration")]
         public async void GenerateAllPagesAndFeatures(string projectType, string framework, string language)
@@ -182,11 +138,6 @@ namespace Microsoft.Templates.Test
         public static IEnumerable<object[]> GetProjectTemplates()
         {
             return GenerationFixture.GetProjectTemplates();
-        }
-
-        public static IEnumerable<object[]> GetPageAndFeatureTemplates()
-        {
-            return GenerationFixture.GetPageAndFeatureTemplates();
         }
     }
 }
