+++ code/test/Templates.Test/Templates.Test.csproj
@@ -117,6 +117,7 @@
     <Compile Include="BaseTestContextProvider.cs" />
     <Compile Include="CodeStyleEnforcementTests.cs" />
     <Compile Include="GlobalSuppressions.cs" />
+    <Compile Include="SingleItemGenerationTests.cs" />
     <Compile Include="StyleCopCollection.cs" />
     <Compile Include="StyleCopGenerationTestsFixture.cs" />
     <Compile Include="StyleCopProjectGenerationTests.cs" />
